FEAT: Improve fle naming, create keyboard and set the first stage of language switcher
