+++ package.json
@@ -30,6 +30,8 @@
     "eslint-plugin-react-hooks": "^4.6.0",
     "framer-motion": "^12.5.0",
     "google-auth-library": "^9.5.0",
+    "i18next": "^24.2.3",
+    "i18next-browser-languagedetector": "^8.0.4",
     "jsonschema": "^1.5.0",
     "jwt-decode": "^4.0.0",
     "prettier": "^2.8.0",
@@ -37,6 +39,7 @@
     "react": "^19.0.0",
     "react-circular-progressbar": "^2.1.0",
     "react-dom": "^19.0.0",
+    "react-i18next": "^15.4.1",
     "react-router-dom": "^7.3.0",
     "react-scripts": "5.0.1",
     "styled-components": "^6.1.15",
