+++ src/i18n/locales/en/dashboard.json
@@ -1,0 +1,3 @@
+{
+  "title": "Accounts balance"
+}
