+++ src/i18n/locales/es/dashboard.json
@@ -1,0 +1,3 @@
+{
+  "title": "Balance"
+}
