+++ src/index.js
@@ -2,6 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import reportWebVitals from "./reportWebVitals";
 import App from "./App";
+import "./i18n/i18n";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
