+++ src/modules/components/common/ColorSelector.js
