--- src/modules/components/more/MoreContent.js
@@ -1,59 +1,0 @@
-import React, { useState } from "react";
-import { Box } from "@mui/material";
-import {
-  ChipComponent,
-  PaperComponent,
-} from "../../styles/more/MoreContent.styled";
-import SettingsContent from "./settings/SettingsContent";
-import ManageContent from "./manage/ManageContent";
-import CommonHeaderContent from "../common/CommonHeaderContent";
-
-/**
- * MoreContent Component
- * Displays additional settings and options, including a logout button.
- *
- * @returns {React.JSX.Element} A section with selectable options and a logout button.
- */
-const MoreContent = () => {
-  const [moreCurrentPage, setMoreCurrentPage] = useState("Manage");
-  const chipElements = ["Manage", "Follow", "About"];
-
-  return (
-    <>
-      {/* Common Header */}
-      <CommonHeaderContent
-        handleClose={() => {}}
-        title={"More Options"}
-        showSettings={true}
-      />
-
-      {/* ChipContent */}
-      <PaperComponent sx={{ backgroundColor: "secondary.main" }}>
-        <Box display="flex" justifyContent="space-between" gap={1}>
-          {chipElements.map((option) => (
-            <ChipComponent
-              key={option}
-              label={option}
-              onClick={() => setMoreCurrentPage(option)}
-              selected={moreCurrentPage}
-              option={option}
-            />
-          ))}
-        </Box>
-      </PaperComponent>
-
-      {/* SettingsContent */}
-      {moreCurrentPage === "Manage" ? <ManageContent /> : null}
-
-      {/* SettingsContent */}
-      <SettingsContent />
-    </>
-  );
-};
-
-/**
- * MoreContent propTypes
- */
-MoreContent.propTypes = {};
-
-export default MoreContent;
