--- src/modules/language/en_us.json
@@ -1,20 +1,0 @@
-{
-    "title": "Money Map",
-    "languageSettings": {
-        "changeLangTitle": "Change Language:",
-        "langTitles": {
-            "en": "English",
-            "es": "Spanish"
-        }
-    },
-    "notifications": {
-        "login": "You have been logged in successfully.",
-        "logout": "You have been logged out successfully."
-    },
-    "splash": {
-        "buttonText": "Go to Dashboard"
-    },
-    "dashboard": {
-        "logoutButtonText": "Logout"
-    }
-}
