--- src/modules/language/es_es.json
@@ -1,20 +1,0 @@
-{
-    "title": "Money Map",
-    "languageSettings": {
-        "changeLangTitle": "Cambiar Idioma:",
-        "langTitles": {
-            "en": "Inglés",
-            "es": "Español"
-        }
-    },
-    "notifications": {
-        "login": "Usuario autenticado con éxito!",
-        "logout": "La sesión fue cerrada con éxito!"
-    },
-    "splash": {
-        "buttonText": "Ir al Dashboard"
-    },
-    "dashboard": {
-        "logoutButtonText": "Cerrar Sesión"
-    }
-}
